+++ src/main/java/lotto/model/LottoResults.java
@@ -1,6 +1,6 @@
 package lotto.model;
 
-import lotto.enums.LottoRanking;
+import lotto.utils.LottoRanking;
 
 import java.util.Arrays;
 import java.util.HashMap;
